pyproject.toml
[project]
name = "delta-inspect"
version = "0.1.0"
authors = [
    { name = "Franz Wöllert", email = "franz.woellert@heidelberg.com" }
]
description = "A lightweight Python CLI and API toolbox for **inspecting Delta Lake tables**."
requires-python = ">=3.12"
dependencies = [
    "typer",
    "pydantic",
    "polars[pyarrow]",
    "deltalake",
    "rtree"
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
    "mypy",
    "deltalake",
    "ipykernel",
    "nbformat",
    "delta-spark>=3.3.2",
]

[project.scripts]
delta-inspect = "delta_inspect.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
